requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
kafka-python==2.0.2
requests==2.32.3
urllib3==2.3.0
